Weather Cards Have Separate Module CSS
