+++ src/App.css
@@ -8,15 +8,15 @@
   font-family: Urbanist, sans-serif;
   text-align: center;
   background-color: #f2f2f2;
-  width: 100vw;
-  height: 100vh;
+  width: 100%;
+  height: 100%;
 }
 
 input[type="text"] {
   font-size: 1.2em;
   border-radius: 0.2em;
   border: 1px solid #ccc;
-  box-shadow: 0 0 1em #ccc;
+  box-shadow: 0 0 1em black;
 }
 
 ul {
