+++ src/components/WeatherDisplay/WeatherCards/CurrentWeatherCard.js
@@ -1,15 +1,18 @@
-import styles from './WeatherCard.module.css'
+import styles from './CurrentWeatherCard.module.css'
+import React from "react";
 
 export const CurrentWeatherCard = ({data , thunderWeather, weatherCondition} ) => {
 
     return (
-            <div className={styles.weather__condition}>
-                <div className={styles.weather__condition__icon}>
-                    <h2>{data.current.temp_c}°C</h2>
+            <div className={styles.currentDay__card}>
+                <h2>{data?.location.name}</h2>
+                <div className={styles.currentDay__card__icon}>
                     {thunderWeather(data?.current.condition.text.toLowerCase())}
                     {weatherCondition(data?.current.condition.text.toLowerCase(),data?.current.is_day)}
+                    <h2>{data.current.temp_c}°C</h2>
                 </div>
-                <p>Date: {data.location.localtime.split(" ")[0]}</p>
+                <h3>Today</h3>
+                <h3>Date: {data.location.localtime.split(" ")[0]}</h3>
                 <p>Current Time: {data.location.localtime.split(" ")[1]}</p>
                 <p>Condition: {data.current.condition.text}</p>
                 <p>Humidity: {data.current.humidity}%</p>
