+++ src/components/WeatherDisplay/WeatherCards/CurrentWeatherCard.module.css
@@ -1,0 +1,18 @@
+.currentDay__card {
+    margin: 0.5em;
+    padding: 0.5em;
+    font-size: 1.2em;
+    border-radius: 1em;
+    background-color: #fff;
+    color: black;
+    box-shadow: 0 0 0.5em black;
+}
+
+
+.currentDay__card__icon {
+    font-size: 1.3em;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    align-items: center;
+}
