--- src/components/WeatherDisplay/WeatherCards/WeatherCard.module.css
@@ -1,7 +1,0 @@
-.weather__condition__icon {
-    font-size: 1.3em;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-evenly;
-    align-items: center;
-}
